+++ bookmarks.js
@@ -1,7 +1,7 @@
 /**
  * SyncMarks
  *
- * @version 1.2.11
+ * @version 1.2.12
  * @author Offerel
  * @copyright Copyright (c) 2020, Offerel
  * @license GNU General Public License, version 3
@@ -14,7 +14,7 @@ document.addEventListener("DOMContentLoaded", function(event) {
 		bdiv.innerHTML = '';
 		allmarks.forEach(bookmark => {
 			bdiv.appendChild(bookmark);
-			if(bookmark.innerText.toUpperCase().includes(sfilter.toUpperCase())) {
+			if(bookmark.innerText.toUpperCase().includes(sfilter.toUpperCase()) || bookmark.firstChild.attributes.href.nodeValue.toUpperCase().includes(sfilter.toUpperCase())) {
 				bookmark.style.display = 'block';
 				bookmark.style.paddingLeft = '20px';
 			} else {
@@ -70,9 +70,10 @@ document.addEventListener("DOMContentLoaded", function(event) {
 		}
 	});
 
-	document.querySelector("#mngcform input[type='text']").addEventListener('focus', function() {
-		this.select();
-	});
+	if(document.querySelector("#mngcform input[type='text']"))
+		document.querySelector("#mngcform input[type='text']").addEventListener('focus', function() {
+			this.select();
+		});
 
 	document.getElementById("save").addEventListener('click', function(event) {
 		event.preventDefault();
@@ -301,6 +302,10 @@ document.addEventListener("DOMContentLoaded", function(event) {
 		});
 	});
 
+	document.getElementById('fname').addEventListener('input', function() {
+		document.getElementById('fsave').disabled = false;
+	});
+
 	document.getElementById('edtitle').addEventListener('input', function() {
 		document.getElementById('edsave').disabled = false;
 	});
@@ -311,6 +316,27 @@ document.addEventListener("DOMContentLoaded", function(event) {
 
 	document.getElementById('mvfolder').addEventListener('change', function() {
 		document.getElementById('mvsave').disabled = false;
+	});
+
+	document.getElementById('fsave').addEventListener('click', function(e) {
+		e.preventDefault();
+		let xhr = new XMLHttpRequest();
+		let data = 'caction=cfolder&fname=' + document.getElementById('fname').value + '&fbid=' + document.getElementById('fbid').value;
+		xhr.onreadystatechange = function () {
+			if (this.readyState == 4) {
+				if(this.status == 200) {
+					if(this.responseText == 1)
+						location.reload(false);
+					else
+						console.log("There was a problem adding the new folder."); 
+				}
+			}
+		};
+		xhr.open("POST", document.location.href, true);
+		xhr.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
+		xhr.send(data);
+		hideMenu();
+		return false;
 	});
 
 	document.getElementById('edsave').addEventListener('click', function(e) {
@@ -357,7 +383,7 @@ document.addEventListener("DOMContentLoaded", function(event) {
 
 function delClient(element) {
 	let xhr = new XMLHttpRequest();
-	let data = 'adel=true&cido=' + element.srcElement.parentElement.id;
+	let data = 'adel=true&cido=' + element.target.parentElement.id;
 	xhr.onreadystatechange = function () {
 		if (xhr.readyState == 4) {
 			if(xhr.status == 200) {
@@ -374,7 +400,7 @@ function delClient(element) {
 
 function mvClient(element) {
 	let xhr = new XMLHttpRequest();
-	let data = 'arename=true&cido=' + element.srcElement.parentElement.id + '&nname=' + element.srcElement.parentElement.children[0].children['cname'].value;
+	let data = 'arename=true&cido=' + element.target.parentElement.id + '&nname=' + element.target.parentElement.children[0].children['cname'].value;
 	xhr.onreadystatechange = function () {
 		if (xhr.readyState == 4) {
 			if(xhr.status == 200) {
@@ -455,7 +481,6 @@ function hideMenu(){
 	let menu = document.querySelector('.menu');
 	menu.classList.remove('show-menu');
 	menu.style.display = 'none';
-	
 	document.querySelectorAll('.mmenu').forEach(function(item) {item.style.display = 'none'});
 	document.querySelectorAll('.mbmdialog').forEach(function(item) {item.style.display = 'none'});
 }
@@ -465,12 +490,13 @@ function onContextMenu(e){
 	hideMenu();
 	let menu = document.querySelector('.menu');
 	menu.style.display = 'block';
-	document.getElementById('bmid').value = e.srcElement.attributes.id.value;
-	document.getElementById('bmid').title = e.srcElement.attributes.title.value;
+	document.getElementById('bmid').value = e.target.attributes.id.value;
+	document.getElementById('bmid').title = e.target.attributes.title.value;
 	showMenu(e.pageX, e.pageY);
 	document.querySelector('#btnEdit').addEventListener('click', onClick, false);
 	document.querySelector('#btnMove').addEventListener('click', onClick, false);
 	document.querySelector('#btnDelete').addEventListener('click', onClick, false);
+	document.querySelector('#btnFolder').addEventListener('click', onClick, false);
 }
 
 function onClick(e){
@@ -500,7 +526,15 @@ function onClick(e){
 			document.getElementById('bmamove').style.display = 'block';
 			break;
 		case 'btnDelete':
-			delBookmark(document.getElementById('bmid').value, document.getElementById('bmid').title)
+			delBookmark(document.getElementById('bmid').value, document.getElementById('bmid').title);
+			break;
+		case 'btnFolder':
+			hideMenu();
+			document.getElementById('folderf').style.left = xpos;
+			document.getElementById('folderf').style.top = ypos;
+			document.getElementById('folderf').style.display = 'block';
+			document.getElementById('fbid').value = document.getElementById('bmid').value;
+			//mkFolder(document.getElementById('bmid').value);
 			break;
 		default:
 			break;
@@ -533,6 +567,7 @@ function delMessage(message) {
 		if (this.readyState == 4) {
 			if(this.status == 200) {
 				document.querySelector('#'+loop+' .NotiTable .NotiTableBody').innerHTML = this.responseText;
+				document.querySelectorAll('.NotiTableCell .fa-trash-o').forEach(function(element) {element.addEventListener('click', delMessage, false)});
 			}
 		}
 	};